+++ packages/core/drawer/src/index.js
@@ -1,5 +1,11 @@
 // @flow
 
-export { default, DrawerItemTheme } from './components';
+export {
+  default,
+  DrawerItemTheme,
+  DrawerSkeletonHeader,
+  DrawerSkeletonItem,
+  DrawerItemGroup,
+} from './components';
 
 export * from './components/types';