+++ packages/search/global-search/src/components/GlobalQuickSearchWrapper.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import memoizeOne from 'memoize-one';
 import { CancelableEvent } from '@atlaskit/quick-search';
 import HomeQuickSearchContainer from './home/HomeQuickSearchContainer';
 import ConfluenceQuickSearchContainer from './confluence/ConfluenceQuickSearchContainer';
@@ -178,8 +177,6 @@ export default class GlobalQuickSearchWrapper extends React.Component<Props> {
   static defaultProps = {
     logger: DEFAULT_NOOP_LOGGER,
   };
-  // configureSearchClients is a potentially expensive function that we don't want to invoke on re-renders
-  memoizedConfigureSearchClients = memoizeOne(configureSearchClients);
 
   private makeConfig() {
     const config: Partial<Config> = {};
@@ -218,7 +215,7 @@ export default class GlobalQuickSearchWrapper extends React.Component<Props> {
     } else if (this.props.context === 'home') {
       return HomeQuickSearchContainer;
     } else if (this.props.context === 'jira') {
-      return JiraQuickSearchContainer;
+      return JiraQuickSearchContainer as React.ComponentClass<any>;
     } else {
       // fallback to home if nothing specified
       return HomeQuickSearchContainer;
@@ -263,11 +260,12 @@ export default class GlobalQuickSearchWrapper extends React.Component<Props> {
       <MessagesIntlProvider>
         <GlobalSearchPreFetchContext.Consumer>
           {({ prefetchedResults }) => {
-            const searchClients = this.memoizedConfigureSearchClients(
+            const searchClients = configureSearchClients(
               this.props.cloudId,
               this.makeConfig(),
               prefetchedResults,
             );
+
             const {
               linkComponent,
               isSendSearchTermsEnabled,