+++ packages/search/global-search/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @atlaskit/global-search
 
+## 8.4.0
+- [minor] [438313ec7a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/438313ec7a):
+
+  - Improved performance with a number of small changes to reduce re-rendering
+
+## 8.3.1
+- [patch] [40c934a6dd](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/40c934a6dd):
+
+  - added faster search analytics
+
+## 8.3.0
+- [minor] [4d5fb33572](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4d5fb33572):
+
+  - Jira new design for advanced search
+
+## 8.2.3
+- [patch] [1e01ef79e2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e01ef79e2):
+
+  - Put queryVersion into the request for searches
+
+## 8.2.2
+- [patch] [ebc992e51d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ebc992e51d):
+
+  - Adds referral context identifiers to more analytics events.
+
 ## 8.2.1
 - [patch] [4051c8b935](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4051c8b935):